59 support controllers devices that manage various things

+++ project/backend-services/smart-irrigation-backend/domain-services/src/main/java/pt/sensae/services/smart/irrigation/backend/domainservices/garden/model/BoundaryCommandDetails.java
@@ -8,5 +8,5 @@ public class BoundaryCommandDetails {
 
     public Double longitude;
 
-    public Double altitude;
+    public Float altitude;
 }
