+++ project/backend-services/smart-irrigation-backend/domain/src/main/java/pt/sensae/services/smart/irrigation/backend/domain/model/GPSPoint.java
@@ -1,8 +1,9 @@
 package pt.sensae.services.smart.irrigation.backend.domain.model;
 
-import pt.sharespot.iot.core.sensor.data.GPSDataDTO;
 
-public record GPSPoint(Double latitude, Double longitude, Double altitude) {
+import pt.sharespot.iot.core.sensor.data.types.GPSDataDTO;
+
+public record GPSPoint(Double latitude, Double longitude, Float altitude) {
 
     public static GPSPoint from(GPSDataDTO data) {
         return new GPSPoint(data.latitude, data.longitude, data.altitude);
@@ -12,7 +13,7 @@ public record GPSPoint(Double latitude, Double longitude, Double altitude) {
         return new GPSPoint(latitude, longitude, null);
     }
 
-    public static GPSPoint ofLatLongAlt(Double latitude, Double longitude, Double altitude) {
+    public static GPSPoint ofLatLongAlt(Double latitude, Double longitude, Float altitude) {
         return new GPSPoint(latitude, longitude, altitude);
     }
 }
