+++ project/backend-services/smart-irrigation-backend/infrastructure/persistence/postgres/src/main/java/pt/sensae/services/smart/irrigation/backend/infrastructure/persistence/postgres/mapper/device/LedgerMapper.java
@@ -33,7 +33,7 @@ public class LedgerMapper {
     }
 
     public static LedgerEntry daoToModel(LedgerEntryPostgres dao, DeviceRecords records) {
-        var aFloat = dao.altitude != null ? Double.parseDouble(dao.altitude) : null;
+        var aFloat = dao.altitude != null ? Float.parseFloat(dao.altitude) : null;
 
         var gpsPoint = new GPSPoint(Double.parseDouble(dao.latitude), Double.parseDouble(dao.longitude), aFloat);
         var deviceContent = new DeviceContent(DeviceName.of(dao.deviceName), records, gpsPoint);
